Split out VerificationServiceClass tests
